+++ lib/utils.d.ts
@@ -7,7 +7,7 @@ export declare function pxRatio(): number;
 export declare function getLabelText(label: string | IWidgetLabel, required?: boolean): string;
 export declare function getAltLabel({ alt }: IWidgetLabel): m.Vnode<any, any> | null;
 export declare function imgSrc(path: string, dataUrl?: string): string;
-export declare function getDisplayLabel(label?: string | IWidgetLabel): m.Vnode<any, any> | null;
+export declare function getDisplayLabel(label?: string | IWidgetLabel, labelCls?: string): m.Vnode<any, any> | null;
 export declare function getLabel(id: string, uiClass: IWidgetClasses, label?: string | IWidgetLabel, required?: boolean): m.Vnode<any, any> | null;
 export declare function labelIcon(label: IWidgetLabel): ("" | m.Vnode<any, any> | undefined)[];
 export declare function setValue(val: TPropStream): ({ target: { value } }: {