+++ release-notes.md
@@ -1,5 +1,20 @@
 # Release Notes
 
+## 1.7.0
+
+- Checkbox, Toggle, Radio widgets have unified implementation for label/icon "inner" parts
+- Added `selection...` config options for Checkbox, Toggle, and Radio widgets
+  - `...Layout` - list of parts for "inner part" of widget - `label`, `icon`, `on`, `off`
+  - `...OnLabel`
+  - `...OffLabel`
+  - `...OnActive` - Class applied to "on" part when "checked"
+  - `...OnInactive` - Class applied to "on" part when "unchecked"
+  - `...OffActive`
+  - `...OffInactive`
+- BREAKING: Check/Toggle `options` support removed, use config `selectionLayout`
+- BREAKING: Config `toggleFormat` support removed, use config `selectionLayout` and appropriate on/off icons
+- BREAKING: Radio `labelSide` support removed, use config `selectionLayout`
+
 ## 1.6.1
 
 - Support `badgePosition` in currency/percentage widgets, "left"/"right" options locate symbol relative to input